app/requirements.txt
# Flask
Flask
Gunicorn
SQLAlchemy
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
Flask-Migrate
psycopg2-binary
python-dotenv